Implement `QnameRef`, the QName with additional metadata
Some methods of `QnameStr` searches a colon linearly.
However, they are inefficient when it is planned to use them, because
the colon position is known at the validation time (and is discarded
when only `QnameStr` is returned).

`QnameRef` contains the prefix length (i.e. the colon position) in
additon to the `&QnameStr`, and it makes getting prefix and local part
more efficient.